% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverDownscalingAlgorithm.R
\name{addCellIDs}
\alias{addCellIDs}
\title{Add cell IDs to a data frame}
\usage{
addCellIDs(LC_df, ID_column_name)
}
\arguments{
\item{LC_df}{Data frame of land cover map. Must have a minimum of two
columns. First column must be called 'x' and contain x-coordinates of the
map. Second column must be name 'y' and contain y-coordinates of the map.}

\item{ID_column_name}{Character string giving the name of the column that
will be added to the data frame.}
}
\value{
The input data frame with an added column called 'ID_column_name'
  that contains an identification number for each grid cell.
}
\description{
Sorts a data frame by x-coordinates, followed by y-coordinates. Adds a
  column with a user-defined name to the data frame containing an
  identification number for each grid cell. A grid cell corresponds to one
  row in the data frame.
}
