% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landUseReconciliation.R
\name{adjustCoarseLandUseAreas}
\alias{adjustCoarseLandUseAreas}
\title{Adjust land-use areas for each coarse grid cell by the area of fine-scale
  grid cells}
\usage{
adjustCoarseLandUseAreas(
  coarse_scale_land_use_areas,
  fine_scale_areas,
  coarse_cell_area
)
}
\arguments{
\item{coarse_scale_land_use_areas}{A vector containing values of a single
land-use category in all coarse-scale grid cells.}

\item{fine_scale_areas}{A vector containing the area of fine-scale cells
within each coarse-scale grid cells.}

\item{coarse_cell_area}{The area of a grid cell in the
`coarse_scale_raster`.}
}
\value{
A vector of adjusted land-use areas.
}
\description{
Adjusts the areas in coarse-scale grid cells for a single land-use according
  to the area of fine-scale cells assigned to the coarse cell.
}
\details{
Implements equation 1 from Le Page et al. (2016) to adjust the area of a
  given land-use within a coarse-scale grid cell.
}
