% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landUseReconciliation.R
\name{adjustCoarseLCAreas}
\alias{adjustCoarseLCAreas}
\title{Adjust land cover areas for each coarse grid cell by the area of reference
  map grid cells}
\usage{
adjustCoarseLCAreas(
  coarse_scale_land_use_areas,
  ref_map_areas,
  coarse_cell_area
)
}
\arguments{
\item{coarse_scale_land_use_areas}{Vector containing values of a single land
cover type in all coarse-scale grid cells.}

\item{ref_map_areas}{Vector containing the area of reference map cells within
each coarse-scale grid cell.}

\item{coarse_cell_area}{Area of a coarse-scale grid cell.}
}
\value{
Vector of adjusted land cover deltas.
}
\description{
Adjusts the land cover delta value for a single land cover type according to
  the area of reference map cells assigned to the coarse cell.
}
\details{
Implements equation 1 from Le Page et al. (2016) to adjust the area of a
  given land-use within a coarse-scale grid cell.
}
