% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landUseReconciliation.R
\name{aggregateLandUseTypeInOneCell}
\alias{aggregateLandUseTypeInOneCell}
\title{Aggregate a single land-use type in one grid cell}
\usage{
aggregateLandUseTypeInOneCell(grid_cell, agg_rules)
}
\arguments{
\item{grid_cell}{A single row from the `adj_coarse_scale_df` containing
adjusted land-use areas for one coarse-scale grid cell.}

\item{agg_rules}{One column from the `final_land_use_types` matrix, which
gives the rules for aggregating coarse-scale land-use types to a single
fine-scale land-use type.}
}
\value{
The area of the fine-scale land-use type in the given grid cell.
}
\description{
Aggregate a single land-use type in one grid cell
}
