% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverReconciliation.R
\name{aggregateLCClassInOneCell}
\alias{aggregateLCClassInOneCell}
\title{Aggregate a single land cover class in one grid cell}
\usage{
aggregateLCClassInOneCell(grid_cell, agg_rules)
}
\arguments{
\item{grid_cell}{Single row from a data frame containing adjusted land cover
deltas for one coarse-scale grid cell.}

\item{agg_rules}{One column from the `final_LC_classes` matrix, which gives
rules for aggregating coarse-scale land cover classes to a single reference
map land cover class.}
}
\value{
Area of the reference map land cover class in the given grid cell.
}
\description{
Aggregate a single land cover class in one grid cell
}
