% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverDownscalingAlgorithm.R
\name{aggregateLCMap}
\alias{aggregateLCMap}
\title{Aggregate a fine-scale map to the same resolution as a coarser-scale map}
\usage{
aggregateLCMap(ref_map, LC_deltas, ref_map_LC_classes)
}
\arguments{
\item{ref_map}{Data frame of a fine-scale map, which must contain 'x', 'y',
and 'coarse_ID'. The 'coarse_ID' column specifies the closest coarse-scale
grid cell to each fine-scale grid cell. This column can be generated using
the `assignRefMap` function.}

\item{LC_deltas}{Data frame of a coarse-scale map, which must have 'x', 'y',
and 'coarse_ID' columns.}

\item{ref_map_LC_classes}{Vector of land cover types in the reference map. For
an area-based reference map, all land cover types should be column names in
the reference map.}
}
\description{
Aggregate a fine-scale map to the same resolution as a coarser-scale map
}
