% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverReconciliation.R
\name{aggregateLCTypeInOneCell}
\alias{aggregateLCTypeInOneCell}
\title{Aggregate a single land cover type in one grid cell}
\usage{
aggregateLCTypeInOneCell(grid_cell, agg_rules)
}
\arguments{
\item{grid_cell}{Single row from `adj_LC_deltas` containing adjusted land
cover deltas for one coarse-scale grid cell.}

\item{agg_rules}{One column from the `final_LC_types` matrix, which gives
rules for aggregating coarse-scale land cover types to a single reference
map land cover type.}
}
\value{
Area of the reference map land cover type in the given grid cell.
}
\description{
Aggregate a single land cover type in one grid cell
}
