% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landUseReconciliation.R
\name{aggregateToFinalLandUseTypes}
\alias{aggregateToFinalLandUseTypes}
\title{Aggregate land-use types}
\usage{
aggregateToFinalLandUseTypes(final_land_use_types, adj_coarse_scale_df)
}
\arguments{
\item{final_land_use_types}{A matrix containing the fraction of each
coarse-scale land-use type that contributes to each fine-scale land-use
type. Columns should contain fine-scale land-use categories, and rows are
the coarse-scale land-use categories. Each cell should contain the
proportion of the coarse-scale land-use category that contributes to the
fine-scale category in the output map.}

\item{adj_coarse_scale_df}{A data frame of land-use areas within
coarse-scale cells, with every area adjusted for the area of fine-scale
cells associated with that coarse grid cell. This data frame is output from
`reconcileLandUseAreas`.}
}
\value{
A data frame of land-use areas in coarse-scale cells with land-use
  types aggregated to the fine-scale land-use types.
}
\description{
Aggregate land-use types from the coarse-scale raster to the land-use types
  in the fine-scale raster.
}
