% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverReconciliation.R
\name{aggregateToFinalLCClasses}
\alias{aggregateToFinalLCClasses}
\title{Aggregate land cover classes}
\usage{
aggregateToFinalLCClasses(final_LC_classes, adj_LC_deltas)
}
\arguments{
\item{final_LC_classes}{A matrix containing the fraction of each coarse-scale
land cover class that contributes to each reference map land cover class.
Columns should contain reference map land cover classes, and rows are the
coarse-scale land cover classes. Each cell should contain the proportion of
the coarse-scale land cover class that contributes to the fine-scale class in
the output map.}

\item{adj_LC_deltas}{`LCMap` object containing land cover change areas
adjusted for the area of reference map cells associated with each coarse
grid cell. This object is output from `reconcileLCDeltas`.}
}
\value{
`LCMap` object of land cover change in `adj_LC_deltas` where land
  cover classes have been aggregated to the reference map land cover classes.
}
\description{
Aggregate land cover classes from the land cover deltas data frame to the
  land cover classes in the reference map.
}
