% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landUseReconciliation.R
\name{aggregateToFinalLCTypes}
\alias{aggregateToFinalLCTypes}
\title{Aggregate land cover types}
\usage{
aggregateToFinalLCTypes(final_LC_types, adj_LC_deltas)
}
\arguments{
\item{final_LC_types}{A matrix containing the fraction of each coarse-scale
land cover type that contributes to each reference map land cover type.
Columns should contain reference map land cover types, and rows are the
coarse-scale land cover types. Each cell should contain the proportion of
the coarse-scale land cover type that contributes to the fine-scale type in
the output map.}

\item{adj_LC_deltas}{Data frame of land cover deltas, with every value
adjusted for the area of reference map cells associated with that coarse
grid cell. This data frame is output from `reconcileLCDeltas`.}
}
\value{
Data frame of land cover deltas in coarse-scale cells with land cover
  types aggregated to the reference map land cover types.
}
\description{
Aggregate land cover types from the land cover deltas data frame to the
  land cover types in the reference map.
}
