% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{allocateLCs}
\alias{allocateLCs}
\title{Allocate land cover change to reference map}
\usage{
allocateLCs(LC_allocation_params, allocation_type = "intensify")
}
\arguments{
\item{LC_allocation_params}{`LCAllocationParams` object containing all the
parameters required to allocate land cover change to a reference map for a
single timestep.}

\item{allocation_type}{The type of land cover allocation. Can be one of
`intensify` or `expand`, which mean that land covers will be allocated to
cells where they do or do not already occur, respectively.}
}
\value{
`LCAllocationParams` object with an updated reference map after land
  cover change allocation and a data frame with any unallocated land cover
  change areas.
}
\description{
Allocates land cover change (delta) values to the given reference map. Land
  cover classes can be allocated to cells where they already exist
  (intensified) or to cells where they do not exist (expanded). The type of
  land cover allocation can be set with the `allocation_type` argument.
}
