% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landUseReconciliation.R
\name{assignFineScaleCells}
\alias{assignFineScaleCells}
\title{Assign fine-scale cells to coarse-scale grid cells}
\usage{
assignFineScaleCells(fine_scale_raster, coarse_scale_raster)
}
\arguments{
\item{fine_scale_raster}{A raster of fine-scale cells.}

\item{coarse_scale_raster}{A raster of coarse-scale cells to which to assign
the fine-scale cells. Must have be on same projection as the fine-scale
cells.}
}
\value{
A list, where each element corresponds to one coarse-scale grid cell.
  Within each element is a data frame with the cell numbers of fine-scale
  cells and their values.
}
\description{
Assigns fine-scale cells on an initial land-use map to coarse-scale grid
cells from a land-use model.
}
