% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landUseReconciliation.R
\name{calculateFineScaleAreaForCoarseCell}
\alias{calculateFineScaleAreaForCoarseCell}
\title{Calculate area of fine-scale cells assigned to a coarse grid cell}
\usage{
calculateFineScaleAreaForCoarseCell(
  coarse_cell,
  fine_scale_df_assigned,
  fine_cell_area
)
}
\arguments{
\item{coarse_cell}{A row containing land-use areas and ID number for a single
coarse grid cell.}

\item{fine_scale_df_assigned}{A fine-scale cells data frame with a column
containing the ID of the coarse grid cell to which each fine-scale cell
belongs. This is output by the `assignFineScaleCells` function.}

\item{fine_cell_area}{The area of a grid cell in the `fine_scale_raster`.}
}
\value{
The area of fine-scale cells that are assigned to the given coarse
  grid cell.
}
\description{
Calculates the area of fine-scale cells that are assigned to a single coarse
  grid cell
}
