% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelDensities.R
\name{calculateKernelDensitiesForOneCell}
\alias{calculateKernelDensitiesForOneCell}
\title{Calculate kernel densities for one land cover class in one cell}
\usage{
calculateKernelDensitiesForOneCell(
  grid_cell,
  assigned_ref_map,
  LC_class,
  cell_x_dist,
  cell_y_dist
)
}
\arguments{
\item{grid_cell}{Single row from reference map data frame with x and y values
and the area of each land cover type in that cell.}

\item{assigned_ref_map}{Data frame of the reference map which should be the
scale at which you want to downscale the land cover data. The data frame
must have a column called `coarse_ID` which assigns each cell in the
reference map to the nearest coarse-scale cell.}

\item{LC_class}{Single land cover class for which to calculate kernel
densities. Must be a column name in `ref_map_cells_df` and
`assigned_ref_map`.}

\item{cell_x_dist}{Radius of the cell neighbourhood in the x-direction.}

\item{cell_y_dist}{Radius of the cell neighbourhood in the y-direction.}
}
\value{
The kernel density value for the user-specified land cover class and
  grid cell.
}
\description{
Calculates the kernel density values for one land cover class in a single
  grid cell.
}
