% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landUseAllocation.R
\name{calculateLandUseDeltas}
\alias{calculateLandUseDeltas}
\title{Calculate land-use delta values}
\usage{
calculateLandUseDeltas(
  fine_scale_df_with_IDs,
  agg_adj_coarse_scale_df,
  land_use_types
)
}
\arguments{
\item{fine_scale_df_with_IDs}{A data frame of fine-scale cells including the
the ID of the coarse-scale cell that each fine-scale cell is assigned to.}

\item{agg_adj_coarse_scale_df}{Coarse-scale land-use data frame containing
the aggregated and adjusted land-use areas.}

\item{land_use_types}{The names of columns containing land-use values in the
`final_scale_df_with_IDs` data frame.}
}
\value{
A data frame with the delta value for each land-use type in each
  coarse-scale grid cell.
}
\description{
Calculates the change in each land-use type (delta) in every coarse-scale
  cell between two time steps. The fine-scale data frame is used as the
  baseline for the previous timestep, and the coarse-scale data frame gives
  the new land-use areas for the subsequent timestep.
}
