% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landUseAllocation.R
\name{calculateLandUseDeltasForOneCell}
\alias{calculateLandUseDeltasForOneCell}
\title{Calculate land-use delta values for one grid cell}
\usage{
calculateLandUseDeltasForOneCell(
  coarse_grid_cell,
  fine_scale_df_with_IDs,
  land_use_types
)
}
\arguments{
\item{coarse_grid_cell}{The reconciled land-use areas within a single
coarse-scale grid cell.}

\item{fine_scale_df_with_IDs}{A data frame of fine-scale cells including the
the ID of the coarse-scale cell that each fine-scale cell is assigned to.}

\item{land_use_types}{The names of columns containing land-use values in the
`final_scale_df_with_IDs` data frame.}
}
\value{
A one-row data frame giving the delta value for each land-use type in
  the given grid cell.
}
\description{
Calculates the change in land-use types in a single coarse-scale grid cell
  between two time steps.
}
