% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landUseAllocation.R
\name{calculateLandUseDeltasForOneCellAndLandUseType}
\alias{calculateLandUseDeltasForOneCellAndLandUseType}
\title{Calculate area of one land-use type in a coarse-scale cell}
\usage{
calculateLandUseDeltasForOneCellAndLandUseType(
  matched_fine_scale_cells_land_use_areas
)
}
\arguments{
\item{matched_fine_scale_cells_land_use_areas}{Vector with the area of one
land-use type in each fine-scale cells allocated to a single coarse-scale
cell.}
}
\value{
A single value giving the area of one land-use type in all fine-scale
  cells allocated to a single coarse-scale cell.
}
\description{
Sums the area of one land-use type in all fine-scale cells allocated to a
  single coarse-scale cell.
}
