% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverDFManipulation.R
\name{calculateLandUseFraction}
\alias{calculateLandUseFraction}
\title{Calculate land-use fraction}
\usage{
calculateLandUseFraction(land_use_area, total_area)
}
\arguments{
\item{land_use_area}{The area of a specific land-use, for example cropland,
within a grid cell.}

\item{total_area}{The total area of the grid cell.}
}
\value{
The fraction of the area within the cell that is covered by the
  given land-use.
}
\description{
Calculates land-use fraction when given the area of a specific land-use
  within a cell and the total area of the cell.
}
