% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverPreProcessing.R
\name{calculateLCDeltas}
\alias{calculateLCDeltas}
\title{Calculate land-use delta values}
\usage{
calculateLCDeltas(coarse_scale_df_time_1, coarse_scale_df_time_2, LC_types)
}
\arguments{
\item{coarse_scale_df_time_1}{Coarse-scale land cover data frame.}

\item{coarse_scale_df_time_2}{Second coarse-scale land cover data frame from
the subsequent timestep.}

\item{LC_types}{}
}
\value{
A data frame with the delta value for each land cover type between
  the two input timesteps.
}
\description{
Calculates the change in each land cover type (delta) between two timesteps
  of PLUMv2 output.
}
