% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelDensities.R
\name{calculateXYKernelDistances}
\alias{calculateXYKernelDistances}
\title{Calculates the x- and y- distances for the kernel density radius}
\usage{
calculateXYKernelDistances(ref_map_cell_resolution, kernel_radius)
}
\arguments{
\item{ref_map_cell_resolution}{Resolution of one cell in the reference map,
in the form `c(x, y)`.}

\item{kernel_radius}{Radius of cells to include in the kernel density
calculation. A value of 1 means that the neighbour cells used to calculate
kernel density will be 1 cell in every direction around the focal cell.
Defaults to 1.}
}
\value{
Vector in the format `c(x, y)` that gives the radius of
  the cell neighbourhood in the x- and y- directions.
}
\description{
Calculates the x- and y- distances for the kernel density radius
}
