% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAreaChecks.R
\name{checkForUnallocatedLandCover}
\alias{checkForUnallocatedLandCover}
\title{Check for unallocated land cover change after land cover allocation}
\usage{
checkForUnallocatedLandCover(grid_cell, LC_types)
}
\arguments{
\item{grid_cell}{Single row from a land cover data frame as a named vector,
corresponding to one grid cell in a map.}

\item{LC_types}{Vector of land cover types in the grid cell. Land cover types
should all be names of the `grid_cell` vector.}
}
\value{
Invisible 'NULL'. Throws a warning if a grid cell contains
  unallocated land cover change and prints the grid cell to the console.
}
\description{
Checks whether there is any unallocated land cover change after all three
  steps of the land cover allocation algorithm (intensification, expansion
  and second round of intensification). Unallocated land cover change is
  defined as remaining land cover change which is more than 0.01\% of the
  total area of the coarse-scale grid cell.
}
