% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAreaChecks.R
\name{checkLandCoverAreasInOneCell}
\alias{checkLandCoverAreasInOneCell}
\title{Check if sum of land cover area in one grid cell equals the expected total
  area}
\usage{
checkLandCoverAreasInOneCell(grid_cell, total_area, LC_types)
}
\arguments{
\item{grid_cell}{Single row from a land cover data frame as a named vector,
corresponding to one grid cell in a map.}

\item{total_area}{The expected total area in the given grid cell.}

\item{LC_types}{Vector of land cover types in the grid cell. Land cover types
should all be names of the `grid_cell` vector.}
}
\value{
Invisible `NULL`. Throws a warning if the sum of land cover in the
  grid cell is > 0.1\% different to the expected total area of land in the
  cell.
}
\description{
Checks if the sum of all land cover types in one grid cell is within 0.1\% of
  the expected total area of that grid cell.
}
