% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRasters.R
\name{convertFractionsToAreas}
\alias{convertFractionsToAreas}
\title{Convert raster with fractions of land-use to areas of land-use}
\usage{
convertFractionsToAreas(fractions_raster, cell_area)
}
\arguments{
\item{fractions_raster}{A `raster` or `rasterStack` where each grid cell
contains the fraction of area covered by a land-use category.}

\item{cell_area}{The area of a grid cell, which should be the same for every
cell in the raster.}
}
\value{
A `raster` or `rasterStack` object where each grid cell contains the
  area covered by a land-use category.
}
\description{
Takes a raster where each grid cell contains the fraction of a land-use
  category, and converts the value in each grid cell to the area taken up by
  that category.
}
