% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverDownscalingAlgorithm.R
\name{createOutputDir}
\alias{createOutputDir}
\title{Create the output directory for land cover downscaling}
\usage{
createOutputDir(output_dir_path)
}
\arguments{
\item{output_dir_path}{Character, path to the directory in which to saved the
downscaled LULC map files. The directory will be created by \code{\link[=downscaleLC]{downscaleLC()}}
if it does not already exist.}
}
\value{
Creates an output directory with the path given in \code{output_dir_path}.
}
\description{
Create the output directory for land cover downscaling
}
