% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverDownscalingAlgorithm.R
\name{downscaleLC}
\alias{downscaleLC}
\title{Downscale land cover data}
\usage{
downscaleLC(
  ref_map_file_name,
  LC_deltas_file_list,
  LC_deltas_classes,
  ref_map_type = "areas",
  ref_map_LC_classes,
  cell_size_unit = "m",
  match_LC_classes,
  kernel_radius,
  simulation_type = "deterministic",
  fuzzy_multiplier = 1,
  discrete_output_map = FALSE,
  random_seed = as.numeric(Sys.time()),
  output_file_prefix,
  output_dir_path
)
}
\arguments{
\item{ref_map_file_name}{Path of file name for the reference map for
downscaling. The reference map file should be in tab-separated format with
the extension `.txt`. The resolution of the reference map will determine
the resolution of the output downscaled land cover maps.}

\item{LC_deltas_file_list}{List of file names for land cover change (delta).
The algorithm expects one land cover change file per time step. All land
cover change files must be in tab-separated format with the `.txt` file
extension. The list must be ordered by time, with the first time step being
the first file in the list, the second time step the second file, and so
on. Land cover change files can be generated using the `calculateLCDeltas`
function from this package.}

\item{LC_deltas_classes}{Vector of land cover classes in the land cover change
maps. Each land cover class should be a column name in every land cover
change file.}

\item{ref_map_type}{Specifies whether the reference map is discrete (contains
one land cover class per cell) or area-based (provides the area of each
land cover class in each cell). Must be one of "areas" or "discrete".}

\item{ref_map_LC_classes}{Vector of land cover types in the reference map. For
an area-based reference map, all land cover types should be column names in
the reference map.}

\item{cell_size_unit}{The unit that cell areas should be calculated in. Must
be one of "km", "m", or "ha".}

\item{match_LC_classes}{A matrix containing the fraction of each coarse-scale
land cover class that contributes to each reference map land cover class.
Columns should contain reference map land cover classes, and rows are the
coarse-scale land cover classes. Each cell should contain the proportion of
the coarse-scale land cover class that contributes to the fine-scale class in
the output map.}

\item{kernel_radius}{Radius of cells to include in the kernel density
calculation. A value of 1 means that the neighbour cells used to calculate
kernel density will be 1 cell in every direction around the focal cell.
Defaults to 1.}

\item{simulation_type}{Specifies the method of land-use change allocation to
use. Can be 'deterministic', 'fuzzy', or 'null_model'. 'deterministic is
a deterministic simulation where land-use change is only randomly allocated
if a cell has a kernel density of 0. The fuzzy' option gives a stochastic
simulation where each kernel density is modified by adding a modifier that
is drawn from a Normal distribution with mean of 0 and standard deviation
equal to the standard deviation of kernel density values in cells available
for land-use allocation. The 'null_model' option allocations land-use
change entirely randomly.}

\item{fuzzy_multiplier}{A value by which the standard deviation of the Normal
distribution used to draw modifiers in the 'fuzzy' method will be
multiplied. Specifying a value of 2 would mean the Normal distribution has
a mean of 0 and standard deviation of 2 times the standard deviation of the
kernel density values in cells available for land-use change allocation.}

\item{discrete_output_map}{Output discrete land cover as well as area-based
land cover. Default is `FALSE`.}

\item{random_seed}{Numeric random seed for ordering fine-scale cells with a
kernel density value of zero during land cover allocation. Defaults to the
date and time when the function is called.}

\item{output_file_prefix}{Prefix for output downscaled land cover map files.}

\item{output_dir_path}{Path to directory in which to save the downscaled
land cover map files. Will be created if it does not already exist.}
}
\value{
TIFF files containing downscaled land-use maps. If
  `discrete_output_map = TRUE` then maps with both the area of each land-use
  class per cell and the largest land-use class per cell will be generated.
}
\description{
Downscales land cover by applying land cover change from coarse-scale input
  maps to a fine-scale reference map. The method used for downscaling land
  cover is based on that in West et al. (2014) and Le Page et al. (2016).
  Briefly, all fine-scale reference map cells are assigned to a single
  coarse-scale cell using a nearest neighbour approach. Kernel density values
  are calculated for each fine-scale cell and land cover type based on the
  amount of each land cover type in neighbouring cells. The kernel density
  values are used during the land cover allocation process to determine how
  much of a land cover type should be allocated to each grid cell; cells with
  higher kernel density values will be allocated more land cover. Land cover
  change is allocated in three steps: intensification, expansion, followed by
  a second round of intensification.
}
\details{
Land cover allocation is performed in three steps: a first round of
  intensification, where land cover is allocated to cells where it already
  occurs; a round of expansion, where land cover is allocated to cells where
  it does not exist; and a final round of intensification to ensure that all
  land cover change has been allocated. The `intensification_ratio` parameter
  specifies how much land cover the algorithm should attempt to allocate via
  intensification versus expansion. An intensification ratio of 0.8 gives a
  target of 80\% of land cover to be allocated via intensification and 20\%
  by expansion.
}
