% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRasters.R
\name{fillLandUseMatrix}
\alias{fillLandUseMatrix}
\title{Fill land-use matrix with land-use fractions.}
\usage{
fillLandUseMatrix(land_use_df, land_use_type, land_use_matrix)
}
\arguments{
\item{land_use_df}{Data frame of land-use output from one timestep of a
PLUMv2 run.}

\item{land_use_type}{The land-use type for which to generate a raster. Must
be one of the column names in `land_use_df`. Default is "cropland".}

\item{land_use_matrix}{An empty land-use matrix created by
`set_up_land_use_matrix`.}
}
\value{
A land-use matrix filled with fractions of the given `land_use_type`
  in each grid cell.
}
\description{
Fill land-use matrix with land-use fractions.
}
