% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{generateOutputFilePath}
\alias{generateOutputFilePath}
\title{Get output file path string}
\usage{
generateOutputFilePath(dir_path, file_prefix, time_step)
}
\arguments{
\item{dir_path}{Path to directory in which to save the output file.}

\item{file_prefix}{Prefix for land cover map output file.}

\item{time_step}{Time step for the land cover map.}
}
\value{
String giving the path to an output file. The path is made up of the
  directory path, file prefix, timestep, and the suffix 'DownscaledLC.txt'.
}
\description{
Get output file path string
}
