% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{getActualConversions}
\alias{getActualConversions}
\title{Calculate actual conversion areas for a set of grid cells}
\usage{
getActualConversions(cells_for_allocation, LC_conversion_area)
}
\arguments{
\item{cells_for_allocation}{Data frame of fine-scale cells selected to
receive a given land cover class.}

\item{LC_conversion_area}{Remaining area of land in the coarse-scale cell to
be converted from \code{LC_from_name} (decreasing land cover class) to
\code{LC_to_name} (increasing land cover class).}
}
\value{
Data frame of fine-scale cells with an additional column named
\code{actual_conversion} that contains the actual conversion areas for each grid
cell.
}
\description{
For each grid cell, the actual conversion from a decreasing land cover class
to an increasing land cover class is the minimum of the area of the
decreasing land cover class in that cell and the remaining area of land to
be converted within the coarse-scale cell.
}
