% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{getCellsForAllocationKdNotZero}
\alias{getCellsForAllocationKdNotZero}
\title{Return a data frame of fine-scale cells with kernel density greater than 0}
\usage{
getCellsForAllocationKdNotZero(cells_for_allocation)
}
\arguments{
\item{cells_for_allocation}{Data frame of fine-scale cells selected to
receive a given land cover class.}
}
\value{
Data frame of fine-scale cells with kernel density greater than 0,
  sorted from highest to lowest kernel density.
}
\description{
Return a data frame of fine-scale cells with kernel density greater than 0
}
