% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverDownscalingAlgorithm.R
\name{getDiscreteLC}
\alias{getDiscreteLC}
\title{Get discrete land cover classes from a land cover data frame containing the
  area of each land cover class per grid cell}
\usage{
getDiscreteLC(LC_map, ref_map_LC_types)
}
\arguments{
\item{LC_map}{Land cover data frame with the area of each land cover class
per grid cell.}

\item{ref_map_LC_types}{Vector of land cover types in the reference map. For
an area-based reference map, all land cover types should be column names in
the reference map.}
}
\value{
The land cover data frame with an additional column called
  'Discrete_LC_class' that gives a discrete land cover class for each grid
  cell.
}
\description{
Takes a land cover data frame containing the area of each land cover class
  per grid cell and adds a column with a discrete land cover class for each
  grid cell. The discrete land cover class for a cell is calculated as the
  land cover class with the highest area in that cell.
}
