% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{getKernelDensitiesForAllocation}
\alias{getKernelDensitiesForAllocation}
\title{Get kernel densities for a given land cover type for a list of reference map
  cells}
\usage{
getKernelDensitiesForAllocation(kernel_density_df, LC_name, ref_IDs)
}
\arguments{
\item{kernel_density_df}{Data frame of kernel density values for every cell
and land cover type in the reference map. The kernel density data frame
should have the same dimensions as the reference map data frame.}

\item{LC_name}{Name of land cover type to extract kernel density values for.}

\item{ref_IDs}{Vector of reference map cell IDs for which to extract the
kernel density values.}
}
\value{
Data frame of kernel densities for the given land cover type and
  reference map cells.
}
\description{
Extract the kernel density values for a list of reference map cells for a
  specific land cover type.
}
