% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{getLCConversions}
\alias{getLCConversions}
\title{Get amount of one land cover type converted into another}
\usage{
getLCConversions(
  kernel_density_df,
  LC_to_name,
  cells_for_allocation_ref_IDs,
  LC_to_delta,
  LC_from_delta,
  LC_from_values
)
}
\arguments{
\item{kernel_density_df}{Data frame of kernel density values for every cell
and land cover type in the reference map. The kernel density data frame
should have the same dimensions as the reference map data frame.}

\item{LC_to_name}{Name of the land cover that is increasing.}

\item{cells_for_allocation_ref_IDs}{Vector of ID numbers of reference cells
to which the increasing land cover type will be allocated.}

\item{LC_to_delta}{The land cover change amount (delta) for the increasing
land cover type, which is equivalent to the amount that this land cover
type increased in the coarse-scale cell during one timestep.}

\item{LC_from_delta}{The land cover delta value for the decreasing land cover
type, which is the amount that this land cover type decreased in the
coarse-scale cell during one timestep.}

\item{LC_from_values}{The area of the decreasing land cover type within each
reference map cell specified in `cells_for_allocation_ref_IDs`.}
}
\value{
Two column data frame. First column contains IDs of reference map
  cells, second column is the amount of the increasing land cover type
  allocated to each cell.
}
\description{
Calculates the amount of one land cover type that is converted into a second
  land cover type for each reference map cell assigned to a coarse-scale
  cell.
}
