% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{getLCDeltaForTentativeConversion}
\alias{getLCDeltaForTentativeConversion}
\title{Get land cover delta area for tentative land cover conversions}
\usage{
getLCDeltaForTentativeConversion(LC_inc_delta, LC_dec_delta)
}
\arguments{
\item{LC_inc_delta}{The land cover change amount (delta) for the increasing
land cover type, which is equivalent to the amount that this land cover
type increased in the coarse-scale cell during one timestep.}

\item{LC_dec_delta}{The land cover delta value for the decreasing land cover
type, which is the amount that this land cover type decreased in the
coarse-scale cell during one timestep.}
}
\value{
Value of land cover delta to be used in tentative land cover
  conversions.
}
\description{
Returns land cover change area (delta) to be used in tentative land cover
  conversions, which is the minimum of the absolute increasing and decreasing
  land cover type delta values.
}
