% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{getLCTransitions}
\alias{getLCTransitions}
\title{Get the transition matrix for one cell}
\usage{
getLCTransitions(LC_deltas)
}
\arguments{
\item{LC_deltas}{}
}
\value{
Data frame giving order and magnitude of land cover transitions.
}
\description{
The transition matrix determines the area of land cover change that is
  converted to each land cover class.
}
