% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverDownscalingAlgorithm.R
\name{getMaxLCClassInOneCell}
\alias{getMaxLCClassInOneCell}
\title{Get the land cover class with highest area in one grid cell}
\usage{
getMaxLCClassInOneCell(grid_cell_LC)
}
\arguments{
\item{grid_cell_LC}{Named vector containing area of land cover classes in one
grid cell. Names should correspond to the land cover classes in that grid
cell.}
}
\value{
Character string of the land cover class with the greatest coverage
  in the given grid cell.
}
\description{
Takes a named vector with the area of each land cover class in one grid cell,
  and returns the land cover class with the highest area.
}
