% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{getSortedTransitionPriorities}
\alias{getSortedTransitionPriorities}
\title{Get ordered vector of transition priorities for allocation of one land cover
  class}
\usage{
getSortedTransitionPriorities(transition_priorities, row_number)
}
\arguments{
\item{transition_priorities}{Matrix containing transition priorities for land
cover allocation. Each row of the matrix should give the order in which one
land cover type is converted to others.}

\item{row_number}{Row number of the `transition_priorities` matrix containing
the land use for which you want to return an ordered vector of transition
priorities.}
}
\value{
Sorted vector of land cover transition priorities for the land cover
  in the given row.
}
\description{
Takes a land cover transition priorities matrix and a row number, and returns
  an vector with the order of land cover classes in that row.
}
