% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{LCAllocationChecks}
\alias{LCAllocationChecks}
\title{Check unallocated land cover change and the sum of land cover areas in each
  reference map cell}
\usage{
LCAllocationChecks(final_LCs)
}
\arguments{
\item{finalLCs}{`LCAllocationParams` object containing an updated reference
map and unallocated land cover change areas after land cover allocation for
one timestep.}
}
\value{
Returns nothing. Checks for unallocated land cover change and that
  the sum of land cover areas in each reference map cell is equal to the
  original area of the cell.
}
\description{
Check unallocated land cover change and the sum of land cover areas in each
  reference map cell
}
