% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverClasses.R
\docType{class}
\name{LCAllocationParams-class}
\alias{LCAllocationParams-class}
\title{S4 class to hold parameters required for land cover allocation}
\description{
S4 class to hold parameters required for land cover allocation
}
\section{Slots}{

\describe{
\item{\code{LC_deltas}}{`LCMap` object containing the LC deltas data frame and
associated information.}

\item{\code{ref_map}}{`LCMap` object containing a reference map data frame and
associated information.}

\item{\code{transition_priorities}}{Matrix with the transition priorities for land
cover allocation.}

\item{\code{kernel_radius}}{Value for radius of cells used to calculate kernel
densities for each land cover and grid cell.}

\item{\code{intensification_ratio}}{Value giving the ratio of land cover to be
allocated via intensification.}
}}

