% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverClasses.R
\docType{class}
\name{LCDataClass-class}
\alias{LCDataClass-class}
\title{S4 class to hold coarse-scale and fine-scale land cover data for one timestep}
\description{
An S4 class with slots for the reference map data frame and a data frame of
  coarse-scale land cover delta values.
}
\section{Slots}{

\describe{
\item{\code{ref_map_df}}{Data frame of the reference map, where each row is one
raster cell.}

\item{\code{LC_deltas}}{Data frame of land cover changes between two coarse-scale
timesteps. Must be the same projection as the `ref_map_df`.}

\item{\code{LC_types}}{Character vector with the land-use types. The land-use
types should be the same in each data frame.}
}}

