% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverClasses.R
\docType{class}
\name{LCMap-class}
\alias{LCMap-class}
\title{S4 class to hold a land cover map with associated properties}
\description{
S4 class to hold a land cover map with associated properties
}
\section{Slots}{

\describe{
\item{\code{LC_map}}{Data frame containing a land cover map. First two columns must
contain x- and y-coordinates for the map. Each subsequent column contains
the area of a given land cover class per cell.}

\item{\code{agg_LC_map}}{For fine-scale maps only, contains a copy of `LC_map` that
has been aggregated into coarse-scale grid cells.}

\item{\code{LC_classes}}{Vector of land cover classes in the `LC_map` slot. Must be
the same as land cover class column names in the `LC_map` data frame.}

\item{\code{grid_cell_area}}{Gives the area of each grid cell in the `LC_map` data
frame where an equal-area projection is used.}

\item{\code{resolution}}{Vector with the x- and y-dimensions of each grid cell where
all grid cells have the same resolution. Should be provided in the format
`c(x, y)`.}
}}

