% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverDownscalingAlgorithm.R
\name{loadLCDeltas}
\alias{loadLCDeltas}
\title{Load an LC deltas data frame}
\usage{
loadLCDeltas(
  LC_deltas_file_list,
  timestep,
  LC_deltas_classes,
  LC_deltas_cell_area
)
}
\arguments{
\item{LC_deltas_file_list}{List of file names for land cover change (delta).
The algorithm expects one land cover change file per time step. All land
cover change files must be in tab-separated format with the `.txt` file
extension. The list must be ordered by time, with the first time step being
the first file in the list, the second time step the second file, and so
on. Land cover change files can be generated using the `calculateLCDeltas`
function from this package.}

\item{timestep}{Timestep of land cover change to be downscaled.}

\item{LC_deltas_classes}{Vector of land cover classes in the land cover change
maps. Each land cover class should be a column name in every land cover
change file.}

\item{LC_deltas_cell_area}{Area of a single coarse-scale grid cell.}
}
\value{
An `LCMap` object containing the LC deltas data frame and associated
  information.
}
\description{
Loads an LC deltas data frame for the given timestep.
}
