% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverDownscalingAlgorithm.R
\name{loadRefMap}
\alias{loadRefMap}
\title{Load a reference map into R}
\usage{
loadRefMap(
  timestep,
  ref_map_file_name,
  ref_map_type,
  ref_map_LC_classes,
  ref_map_cell_area,
  ref_map_cell_resolution,
  LC_column_name,
  LC_deltas,
  discrete_output_map,
  output_dir_path,
  output_file_prefix
)
}
\arguments{
\item{timestep}{Timestep of land cover change to be downscaled.}

\item{ref_map_file_name}{Path of file name for the reference map for
downscaling. The reference map file should be in tab-separated format with
the extension `.txt`. The resolution of the reference map will determine
the resolution of the output downscaled land cover maps.}

\item{ref_map_type}{Specifies whether the reference map is discrete (contains
one land cover class per cell) or area-based (provides the area of each
land cover class in each cell). Must be one of "areas" or "discrete".}

\item{ref_map_LC_classes}{Vector of land cover types in the reference map. For
an area-based reference map, all land cover types should be column names in
the reference map.}

\item{ref_map_cell_area}{Area of a single reference map grid cell.}

\item{ref_map_cell_resolution}{Resolution of one cell in the reference map,
in the form `c(x, y)`.}

\item{LC_column_name}{For a discrete reference map, gives the name of the
column containing the land cover class for each grid cell.}

\item{LC_deltas}{An `LCMap` object containing land cover change areas for the
given timestep.}

\item{discrete_output_map}{Output discrete land cover as well as area-based
land cover. Default is `FALSE`.}

\item{output_dir_path}{Path to directory in which to save the downscaled
land cover map files. Will be created if it does not already exist.}

\item{output_file_prefix}{Prefix for output downscaled land cover map files.}
}
\value{
An `LCMap` object containing the reference map data frame for the
  given timestep and associated information.
}
\description{
Load a reference map into R
}
