% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landUseDFManipulation.R
\name{makeLandUseFractionDataFrame}
\alias{makeLandUseFractionDataFrame}
\title{Generate land-use fractions data frame}
\usage{
makeLandUseFractionDataFrame(
  land_use_df,
  land_uses = c("managedForest", "unmanagedForest", "otherNatural", "cropland",
    "pasture", "barren", "urban")
)
}
\arguments{
\item{land_use_df}{Data frame of land-use output from PLUMv2. Must have 'Lon'
and 'Lat' columns.}

\item{land_uses}{A list of land-uses for which to convert areas into
fractions.}
}
\value{
A data frame containing the fraction of each land-use in each grid
  cell from PLUMv2.
}
\description{
Generates a data frame containing the fraction of each land-use within each
  output cell from PLUMv2. The included land-uses are: managed forest, other
  natural, cropland, pasture, barren and urban.
}
