% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRasters.R
\name{makeLandUseRaster}
\alias{makeLandUseRaster}
\title{Make raster of land-use from PLUMv2 land-use data frame.}
\usage{
makeLandUseRaster(
  land_use_df,
  land_use_type = "cropland",
  cell_resolution = c(0.5, 0.5),
  project_string = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs",
  lon_column = "Lon",
  lat_column = "Lat"
)
}
\arguments{
\item{land_use_df}{Data frame of land-use output from one timestep of a
PLUMv2 run.}

\item{land_use_type}{The land-use type for which to generate a raster. Must
be one of the column names in `land_use_df`. Default is "cropland".}

\item{cell_resolution}{Resolution of the land-use data, given as c(x, y) or
c(longitude, latitude).
Default is c(0.5, 0.5).}

\item{project_string}{A string that gives the projection for input to the
`raster` function. Default is "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs".}

\item{lon_column}{The name of the column containing longitude values,
defaults to "Lon".}

\item{lat_column}{The name of the column containing latitude values, defaults
to "Lat".}
}
\value{
A raster of the given land-use type. Default is a raster of 0.5
  degree grid cells, each one containing the fraction of cropland in that
  cell for the PLUMv2 timestep.
}
\description{
Make raster of land-use from PLUMv2 land-use data frame.
}
