% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRasters.R
\name{makeLandUseRasterStack}
\alias{makeLandUseRasterStack}
\title{Make raster stack of land-uses from a PLUMv2 land-use data frame.}
\usage{
makeLandUseRasterStack(
  land_use_df,
  raster_layers = c("suitable_fraction", "pa_fraction", "managedForest_fract",
    "unmanagedForest_fract", "otherNatural_fract", "cropland_fract", "pasture_fract",
    "barren_fract", "urban_fract")
)
}
\arguments{
\item{land_use_df}{Data frame of land-use output from one timestep of a
PLUMv2 run.}

\item{raster_layers}{List of land-uses to include in the raster stack. Must
be one of the column names in `land_use_df`.}
}
\value{
A raster stack with each layer corresponding to one land-use.
}
\description{
Make raster stack of land-uses from a PLUMv2 land-use data frame.
}
