% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{multiplyLCDeltas}
\alias{multiplyLCDeltas}
\title{Multiply land cover change values}
\usage{
multiplyLCDeltas(LC_deltas, multiplication_factor)
}
\arguments{
\item{LC_deltas}{An `LCMap` object containing land cover change areas for the
given timestep.}

\item{multiplication_factor}{Factor by which to multiply the land cover
change values in the `LC_deltas` object.}
}
\value{
`LCMap` object containing land cover change values multiplied by the
  user-specified `multiplication_factor` parameter.
}
\description{
Multiplies land cover change (delta) values by a user-specified multiplication
  factor. This allows the implementation of an intensification versus
  expansion factor that determines the amount of land cover which is
  intensified versus expanded during land cover allocation.
}
