% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverReconciliation.R
\name{processLCDeltas}
\alias{processLCDeltas}
\title{Land cover change processing}
\usage{
processLCDeltas(LC_deltas, ref_map, final_LC_classes)
}
\arguments{
\item{LC_deltas}{An `LCMap` object containing land cover change areas for the
given timestep.}

\item{ref_map}{An `LCMap` object containing the reference map for the current
timestep.}

\item{final_LC_classes}{A matrix containing the fraction of each coarse-scale
land cover class that contributes to each reference map land cover class.
Columns should contain reference map land cover classes, and rows are the
coarse-scale land cover classes. Each cell should contain the proportion of
the coarse-scale land cover class that contributes to the fine-scale class in
the output map.}
}
\value{
`LCMap` object containing land cover change from `LC_deltas` that has
  been adjusted according to areas in the reference map and aggregated to
  final land cover classes.
}
\description{
Reconciles the land cover change values for each coarse-scale cell
  with fine-scale areas and aggregates land cover classes to specified final
  land cover classes.
}
