% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{randomiseDataFrame}
\alias{randomiseDataFrame}
\title{Randomly sorts a data frame}
\usage{
randomiseDataFrame(input_df, random_seed)
}
\arguments{
\item{input_df}{Input data frame to be randomly ordered.}

\item{random_seed}{Numeric random seed for ordering fine-scale cells with a
kernel density value of zero during land cover allocation. Defaults to the
date and time when the function is called.}
}
\value{
Input data frame with randomly sorted rows.
}
\description{
Randomly sorts a data frame
}
