% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelDensities.R
\name{randomiseKernelDensities}
\alias{randomiseKernelDensities}
\title{Randomly sorts a data frame with a 'kernel_density' column}
\usage{
randomiseKernelDensities(kernel_density_df, random_seed)
}
\arguments{
\item{kernel_density_df}{Data frame of grid cells with a column named
`kernel_density` that contains a kernel density for each cell.}

\item{random_seed}{Numeric random seed for ordering fine-scale cells with a
kernel density value of zero during land cover allocation. Defaults to the
date and time when the function is called.}
}
\value{
Input data frame with randomly sorted grid cells.
}
\description{
Randomly sorts a data frame with a 'kernel_density' column
}
