% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landUseClasses.R
\docType{class}
\name{ReconciledLandUses-class}
\alias{ReconciledLandUses-class}
\title{S4 class to hold land-use data frames}
\description{
An S4 class with slots for fine-scale and coarse-scale land-use data frames.
}
\section{Slots}{

\describe{
\item{\code{fine_scale_df}}{A data frame of fine-scale land-uses, where each row is
one raster cell.}

\item{\code{coarse_scale_df}}{A data frame of coarse-scale land-uses, where each row
is one raster cell.}

\item{\code{land_use_types}}{A character vector with the land-use types. The land-use
types should be the same in each data frame.}
}}

