% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landUseReconciliation.R
\name{reconcileLandUseAreas}
\alias{reconcileLandUseAreas}
\title{Reconcile land-use areas}
\usage{
reconcileLandUseAreas(
  coarse_scale_df,
  fine_scale_df_assigned,
  coarse_land_use_types,
  coarse_cell_area,
  fine_cell_area
)
}
\arguments{
\item{coarse_scale_df}{A data frame of coarse-scale cells to which to assign
the fine-scale cells. Must have be on same projection as the
fine-scale cells.}

\item{fine_scale_df_assigned}{A fine-scale cells data frame with a column
containing the ID of the coarse grid cell to which each fine-scale cell
belongs. This is output by the `assignFineScaleCells` function.}

\item{coarse_land_use_types}{Vector of land-use types in the coarse-scale
data frame.}

\item{coarse_cell_area}{The area of a grid cell in the
`coarse_scale_raster`.}

\item{fine_cell_area}{The area of a grid cell in the `fine_scale_raster`.}
}
\value{
A data frame of land-use areas within coarse-scale cells, with every
  area adjusted for the area of fine-scale cells associated with that coarse
  grid cell.
}
\description{
Reconcile land-use areas for each coarse-scale grid cell with the area of
  fine-scale cells assigned to that grid cell.
}
\details{
This function uses equation 1 from the Le Page et al. (2016) paper to
  reconcile coarse- and fine-scale land-use areas.
}
