% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landUseReconciliation.R
\name{reconcileLandUses}
\alias{reconcileLandUses}
\title{Land-use reconciliation}
\usage{
reconcileLandUses(
  coarse_scale_raster,
  fine_scale_raster,
  coarse_cell_area,
  fine_cell_area
)
}
\arguments{
\item{coarse_scale_raster}{A raster of land-use areas at a coarse-scale, i.e.
output from PLUMv2.}

\item{fine_scale_raster}{A raster of land-use areas at the scale to which you
want to downscale the `coarse_scale_raster`. Must be in the same projection
as the `coarse_scale_raster`.}

\item{coarse_cell_area}{The area of a grid cell in the
`coarse_scale_raster`.}

\item{fine_cell_area}{The area of a grid cell in the `fine_scale_raster`.}
}
\value{

}
\description{
Performs the following actions: assigns fine-scale cells to coarse-scale
  cells, reconciles the coarse-scale areas with fine-scale areas, and
  aggregates land-uses.
}
