% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverReconciliation.R
\name{reconcileLCDeltas}
\alias{reconcileLCDeltas}
\title{Reconcile land cover deltas}
\usage{
reconcileLCDeltas(LC_deltas, ref_map)
}
\arguments{
\item{LC_deltas}{An `LCMap` object containing land cover change areas for the
given timestep.}

\item{ref_map}{An `LCMap` object containing the reference map for the current
timestep.}
}
\value{
`LCMap` object of land cover change in `LC_deltas` where every area
  is adjusted for the area of fine-scale cells associated with that coarse
  grid cell.
}
\description{
Reconcile land cover deltas for each coarse-scale grid cell with the area of
  reference map cells assigned to that grid cell.
}
\details{
This function uses equation 1 from the Le Page et al. (2016) paper to
  reconcile coarse- and fine-scale land cover areas.
}
