% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{runLCAllocation}
\alias{runLCAllocation}
\title{Run allocation of land covers to reference map by intensification and
  expansion}
\usage{
runLCAllocation(
  assigned_ref_map,
  kernel_density_df,
  LC_deltas,
  transition_priorities,
  intensification_ratio,
  ref_map_cell_area
)
}
\arguments{
\item{assigned_ref_map}{Data frame of the reference map which should be the
scale at which you want to downscale the land cover data. The data frame
must have a column called `coarse_ID` which assigns each cell in the
reference map to the nearest coarse-scale cell.}

\item{kernel_density_df}{Data frame of kernel density values for every cell
and land cover type in the reference map. The kernel density data frame
should have the same dimensions as the reference map data frame.}

\item{LC_deltas}{Data frame of adjusted and aggregated land cover change
(delta) values from the coarse-scale input map for every cell and land
cover type.}

\item{transition_priorities}{Matrix containing transition priorities for land
cover allocation. Each row of the matrix should give the order in which one
land cover type is converted to others.}

\item{intensification_ratio}{Ratio of land intensification versus land
expansion for the allocation algorithm.}

\item{ref_map_cell_area}{Area of a reference map grid cell.}
}
\value{
New land cover data frame with the given land cover change allocated.
}
\description{
This function will run the allocation of all land cover change (delta) values
  from a coarse-scale map, such as the output from one timestep of a global
  land-use model, to a fine-scale reference map. The method employed is from
  West et al. (2014) and Le Page et al. (2016). Land cover change is
  allocated in three steps: a first round of intensification, meaning land
  cover is allocated to cells where it already occurs; a round of expansion,
  where land cover is allocated to cells where it does not exist; and a final
  round of intensification to make sure all land cover change has been
  allocated.
}
\details{
The amount of land cover change that is allocated by intensification versus
  expansion is determined by the `intensification_ratio` parameter. An
  intensification ratio of 0.8 gives a target of 80 percent of land cover to
  be allocated via intensification and 20 percent by expansion.

  The amount of land cover change for a single land cover type that is
  allocated to a fine-scale reference cell is currently determined by the
  kernel density value of that cell. A kernel density value is a relative
  measure of the amount of a land cover type in neighbouring cells.
}
