% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{saveLandCoverMapAsTable}
\alias{saveLandCoverMapAsTable}
\title{Save land cover maps as .txt files}
\usage{
saveLandCoverMapAsTable(LC_map, file_prefix = "", dir_path = "", time_step)
}
\arguments{
\item{LC_map}{Land cover map data frame.}

\item{file_prefix}{Prefix for land cover map output file.}

\item{dir_path}{Path to directory in which to save the output file.}

\item{time_step}{Time step for the land cover map.}
}
\value{
Tab-delimited text file of the given land cover map.
}
\description{
Saves output land cover maps as tab-delimited text files with the `.txt` file
  extension.
}
