% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverDFManipulation.R
\name{selectCell}
\alias{selectCell}
\title{Select cell from land-use data frame}
\usage{
selectCell(land_use_df, lat, lon)
}
\arguments{
\item{land_use_df}{Data frame of raw land use values from PLUMv2 containing
columns named 'Lat' and 'Lon'.}

\item{lat}{Value of latitude for the cell you want to select.}

\item{lon}{Value of longitude for the cell you want to select.}
}
\value{
A row from the `land_use_df` that contains a single grid cell.
}
\description{
Selects a cell from the `raw_land_use` data frame using latitude and
  longitude values.
}
