% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landUseDFManipulation.R
\name{selectRegion}
\alias{selectRegion}
\title{Select region from land-use data frame}
\usage{
selectRegion(land_use_df, min_lat, max_lat, min_lon, max_lon)
}
\arguments{
\item{land_use_df}{Data frame of raw land use values from PLUMv2 containing
columns named 'Lat' and 'Lon'.}

\item{min_lat}{Minimum latitude value of the region to select.}

\item{max_lat}{Maximum latitude value of the region to select.}

\item{min_lon}{Minimum longitude value of the region to select.}

\item{max_lon}{Maximum value of longitude for the region to select.}
}
\value{
A data frame containing rows with latitude and longitude values
  within the specified limits.
}
\description{
Returns a rectangular region from the `raw_land_use` data frame using
  latitude and longitude values.
}
