% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRasters.R
\name{setUpLandUseMatrix}
\alias{setUpLandUseMatrix}
\title{Set up a matrix for land-use data filled with NA.}
\usage{
setUpLandUseMatrix(
  cell_resolution,
  land_use_min_lat,
  land_use_max_lat,
  land_use_min_lon,
  land_use_max_lon
)
}
\arguments{
\item{cell_resolution}{Resolution of the land-use data, given as c(x, y) or
c(longitude, latitude).
Default is c(0.5, 0.5).}

\item{land_use_min_lat}{The minimum latitude in the `land_use_df`.}

\item{land_use_max_lat}{The maximum latitude in the `land_use_df`.}

\item{land_use_min_lon}{The minimum longitude in the `land_use_df`.}

\item{land_use_max_lon}{The maximum longitude in the `land_use_df`.}
}
\value{
A matrix filled with NAs. Columns will be from minimum longitude to
  maximum longitude with step size `cell_resolution[1]`. Rows will be from
  maximum latitude to minimum latitude with step size `cell_resolution[2]`.
}
\description{
Set up a matrix for land-use data filled with NA.
}
