% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{simulateLandCoverData}
\alias{simulateLandCoverData}
\title{Simulate land cover data}
\usage{
simulateLandCoverData(LC_types, res_x, res_y, number_cells_x, number_cells_y)
}
\arguments{
\item{LC_types}{Vector of land cover types to simulate.}

\item{res_x}{Size of one grid cell in the x-direction.}

\item{res_y}{Size of one grid cell in the y-direction.}

\item{number_cells_x}{Number of cells in the x-direction.}

\item{number_cells_y}{Number of cells in the y-direction.}
}
\value{
Data frame of land cover. The first two columns give the x- and
  y-coordinates of grid cells and the remaining columns contain the simulated
  area of each land cover type.
}
\description{
Simulates land cover data of a given resolution and size using random
  sampling. For each grid cell, the `sample` function is used to sample from
  the number of input land cover types. The cell size determines the number
  of items sampled. The number of times each land cover type occurs in the
  sample is the area of that land cover type in the output map.
}
