% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{sortCellsForAllocation}
\alias{sortCellsForAllocation}
\title{Sort cells before land cover allocation}
\usage{
sortCellsForAllocation(cells_for_allocation, random_seed)
}
\arguments{
\item{cells_for_allocation}{Data frame of fine-scale cells selected to
receive a given land cover class.}

\item{random_seed}{Numeric random seed for ordering fine-scale cells with a
kernel density value of zero during land cover allocation. Defaults to the
date and time when the function is called.}
}
\value{
Data frame of fine-scale cells ordered from highest to lowest kernel
  density. Cells with a kernel density value of zero are randomly sorted at
  the end of the data frame.
}
\description{
Sort cells before land cover allocation
}
