% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{sumLCDeltas}
\alias{sumLCDeltas}
\title{Sum land cover change values}
\usage{
sumLCDeltas(LC_deltas_one, LC_deltas_two, LC_types)
}
\arguments{
\item{LC_deltas_one}{Data frame of land cover change values from a
coarse-scale map.}

\item{LC_deltas_two}{Second data frame of land cover change values from a
coarse-scale map.}

\item{LC_types}{Vector of land cover types in the `LC_deltas` data frame.}
}
\value{
Data frame containing the sum of land cover change from the two input
  land cover change data frames.
}
\description{
Sums the land cover change values in two data frames. The land cover types
  and grid cells must be the same in each data frame.
}
