% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{updateLCDecDelta}
\alias{updateLCDecDelta}
\title{Update land cover change (delta) area of a decreasing land cover type}
\usage{
updateLCDecDelta(LC_dec_delta, total_conversion)
}
\arguments{
\item{LC_dec_delta}{The land cover delta value for the decreasing land cover
type, which is the amount that this land cover type decreased in the
coarse-scale cell during one timestep.}

\item{total_conversion}{The total area of land converted from the given land
cover type.}
}
\value{
The new land cover delta value for one land cover type.
}
\description{
Updates the land cover delta value of a decreasing land cover type, by
  adding the total decrease of that land cover within one coarse-scale cell
  to the initial delta value.
}
