% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{updateLCToDelta}
\alias{updateLCToDelta}
\title{Update land cover change (delta) area of an increasing land cover type}
\usage{
updateLCToDelta(LC_to_delta, total_conversion)
}
\arguments{
\item{LC_to_delta}{The land cover change amount (delta) for the increasing
land cover type, which is equivalent to the amount that this land cover
type increased in the coarse-scale cell during one timestep.}

\item{total_conversion}{The total area of land converted to the given land
cover type.}
}
\value{
The new land cover delta value for one land cover type.
}
\description{
Updates the land cover delta value of an increasing land cover type, by
  subtracting the total increase of that land cover within one coarse-scale
  cell from the initial delta value.
}
