% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{updateOneRefMapCellWithLCConversions}
\alias{updateOneRefMapCellWithLCConversions}
\title{Update one reference map cell with land cover conversion areas}
\usage{
updateOneRefMapCellWithLCConversions(
  ref_map_cell,
  LC_inc_name,
  LC_dec_name,
  actual_conversion
)
}
\arguments{
\item{ref_map_cell}{One row from a reference map data frame, which is
equivalent to one grid cell.}

\item{LC_inc_name}{Name of the land cover that is increasing.}

\item{LC_dec_name}{Name of the land cover that `LC_inc_name` is being
converted to.}

\item{actual_conversion}{The actual area of land cover to be converted from
the increasing land cover (`LC_inc_name`) to the decreasing land cover
(`LC_dec_name`) in this reference map cell.}
}
\value{
A one row data frame with updated land cover areas.
}
\description{
Updates one cell from a reference map with land cover conversion areas from
  the conversion of one land cover type to another.
}
