% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{updateRefMapWithLCConversions}
\alias{updateRefMapWithLCConversions}
\title{Update reference map with land cover conversion values}
\usage{
updateRefMapWithLCConversions(
  ref_map_df_with_IDs,
  LC_conversion_df,
  LC_dec_name,
  LC_inc_name
)
}
\arguments{
\item{ref_map_df_with_IDs}{Reference map data frame with a column containing
the ID of the coarse grid cell to which each reference map cell was
assigned to. This is output from the `assignFineScaleCells` function.}

\item{LC_conversion_df}{Data frame with one column for actual land cover
conversion areas and a second with reference map IDs. Output from the
`getLCConversions` function.}

\item{LC_dec_name}{Name of the land cover that `LC_inc_name` is being
converted to.}

\item{LC_inc_name}{Name of the land cover that is increasing.}
}
\value{
Reference map data frame with updated land cover areas in the cells
  specified in the `LC_conversion_df` data frame.
}
\description{
Updates a reference map with the actual land cover conversion areas for one
  timestep and land cover type.
}
