% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{allocateLCs}
\alias{allocateLCs}
\title{Allocate land cover change to reference map}
\usage{
allocateLCs(
  ref_map_df_with_IDs,
  kernel_density_df,
  LC_deltas,
  transition_priorities,
  allocation_type = "intensify"
)
}
\arguments{
\item{ref_map_df_with_IDs}{Reference map data frame with a column containing
the ID of the coarse grid cell to which each reference map cell was
assigned to. This is output from the `assignFineScaleCells` function.}

\item{kernel_density_df}{Data frame of kernel density values for every cell
and land cover type in the reference map. The kernel density data frame
should have the same dimensions as the reference map data frame.}

\item{LC_deltas}{Data frame of adjusted and aggregated land cover change
(delta) values from the coarse-scale input map for every cell and land
cover type.}

\item{transition_priorities}{Matrix containing transition priorities for land
cover allocation. Each row of the matrix should give the order in which one
land cover type is converted to others.}

\item{allocation_type}{The type of land cover allocation. Can be one of
`intensify` or `expand`, which mean that land covers will be allocated to
cells where they do or do not already occur, respectively.}
}
\value{
`LCDataClass` with a new reference map with land cover change
  allocated, data frame with land cover delta values, and the final land
  cover types in the reference map.
}
\description{
Allocates land cover change (delta) values to the given reference map. Land
  cover types can be allocated to cells where they already exist
  (intensified) or to cells where they do not exist (expanded). The type of
  land cover allocation can be set with the `allocation_type` argument.
}
