% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landUseReconciliation.R
\name{assignFineScaleCells}
\alias{assignFineScaleCells}
\title{Assign fine-scale cells to coarse-scale grid cells}
\usage{
assignFineScaleCells(fine_scale_df, coarse_scale_df)
}
\arguments{
\item{fine_scale_raster}{A data frame of fine-scale cells.}

\item{coarse_scale_raster}{A data frame of coarse-scale cells to which to
assign the fine-scale cells. Must have be on same projection as the fine-scale
cells.}
}
\value{
A copy of the fine-scale cells data frame with an extra column
  containing the ID of the coarse-scale cell to which each fine-scale cell
  belongs.
}
\description{
Assigns fine-scale cells on an initial land-use map to coarse-scale grid
cells from a land-use model using a nearest-neighbour method.
}
