% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverReconciliation.R
\name{assignRefMapCells}
\alias{assignRefMapCells}
\title{Assign reference map cells to coarse-scale grid cells}
\usage{
assignRefMapCells(ref_map_df, LC_deltas)
}
\arguments{
\item{ref_map_df}{Data frame of the reference map, which is the scale at
which you want to downscale the land cover data.}

\item{LC_deltas}{Data frame of land cover changes between two coarse-scale
timesteps. Must be the same projection as `ref_map_df`.}
}
\value{
A copy of the fine-scale cells data frame with an extra column
  containing the ID of the coarse-scale cell to which each fine-scale cell
  belongs.
}
\description{
Assigns fine-scale cells on an reference land cover map to coarse-scale grid
  cells from a land-use model using a nearest-neighbour method.
}
