% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverDownscalingAlgorithm.R
\name{assignRefMapCells}
\alias{assignRefMapCells}
\title{Assign fine resolution cells to a coarse resolution map}
\usage{
assignRefMapCells(ref_map, LC_deltas_coords, LC_deltas_cell_numbers)
}
\description{
Assign fine resolution cells to a coarse resolution map
}
