% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverDownscalingAlgorithm.R
\name{assignRefMapCells}
\alias{assignRefMapCells}
\title{Assign fine resolution grid cells to a coarse resolution map}
\usage{
assignRefMapCells(ref_map, LC_deltas_coords, LC_deltas_cell_numbers)
}
\arguments{
\item{ref_map}{\code{\link[terra:SpatRaster-class]{terra::SpatRaster}} object at fine resolution.}

\item{LC_deltas_coords}{Matrix, coordinates of coarse resolution grid cells.
Can be obtained from a \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} object using the
\code{\link[terra:crds]{terra::crds()}} function.}

\item{LC_deltas_cell_numbers}{Numeric, cell numbers that uniquely identify
each cell in the coarse resolution map. Should correspond to the order of
coordinates in \code{LC_deltas_coords}. Can be obtained from a
\code{\link[terra:SpatRaster-class]{terra::SpatRaster}} object using the \code{\link[terra:cells]{terra::cells()}} function.}
}
\value{
\code{\link[terra:SpatVector-class]{terra::SpatVector}} object of numbered polygons. The number of each
polygon corresponds to the unique identifier of a coarse resolution grid
cell. Each polygon encapsulates a set of fine resolution cells that have
been assigned to that coarse resolution cell.
}
\description{
Assigns grid cells from a fine resolution map to their nearest neighbour grid
cell in a coarse resolution map using the \code{\link[FNN:get.knn]{FNN::get.knnx()}} function.
}
