% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{calculateActualConversion}
\alias{calculateActualConversion}
\title{Calculate actual conversion values}
\usage{
calculateActualConversion(LC_from_values, tentative_conversion)
}
\arguments{
\item{LC_from_values}{Vector of the area of the land cover type to be
decreased that is currently in a set of reference map cells.}

\item{tentative_conversion}{Vector with the tentative land cover conversion
value for each reference map cell.}
}
\value{
Vector with the actual land cover conversion area for the given set
  of reference map cells.
}
\description{
Calculates actual land cover conversion areas for each reference map cell,
  which is the minimum of the tentative conversion value for a cell and the
  amount of the decreasing land cover currently in that cell.
}
