% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelDensities.R
\name{calculateKernelDensities}
\alias{calculateKernelDensities}
\title{Calculate kernel densities}
\usage{
calculateKernelDensities(
  fine_scale_df_with_IDs,
  land_use_types,
  cell_resolution,
  moving_window = 1
)
}
\arguments{
\item{fine_scale_df_with_IDs}{A data frame of fine-scale cells including the
the ID of the coarse-scale cell that each fine-scale cell is assigned to.}

\item{land_use_types}{The names of columns containing land-use values in the
`final_scale_df_with_IDs` data frame.}

\item{cell_resolution}{The resolution of one cell in the landscape. Should be
in the form `c(x, y)`.}

\item{moving_window}{The radius of cells to include in the kernel density
calculation. A value of 1 means that the neighbour cells for calculating
kernel density will be 1 cell in every direction around the focal cell.
Defaults to 1.}
}
\value{
A data frame with the kernel density value for each land-use type in
  each cell.
}
\description{
Calculates kernel densities for all cell and land-use types in a data frame.
}
