% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelDensities.R
\name{calculateKernelDensities}
\alias{calculateKernelDensities}
\title{Calculate kernel densities}
\usage{
calculateKernelDensities(
  ref_map_df_with_IDs,
  ref_map_LC_types,
  cell_resolution,
  moving_window = 1
)
}
\arguments{
\item{ref_map_df_with_IDs}{Reference map data frame with a column containing
the ID of the coarse grid cell to which each reference map cell was
assigned to. This is output from the `assignFineScaleCells` function.}

\item{ref_map_LC_types}{Vector of land cover types in the reference map.}

\item{cell_resolution}{Resolution of one cell in the map, in the form
`c(x, y)`.}

\item{moving_window}{Radius of cells to include in the kernel density
calculation. A value of 1 means that the neighbour cells used to calculate
kernel density will be 1 cell in every direction around the focal cell.
Defaults to 1.}
}
\value{
Data frame with the kernel density value for each land cover type in
  each cell.
}
\description{
Calculates kernel densities for all cell and land cover types in a reference
  map data frame.
}
