% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelDensities.R
\name{calculateKernelDensities}
\alias{calculateKernelDensities}
\title{Calculate kernel densities}
\usage{
calculateKernelDensities(
  assigned_ref_map,
  ref_map_LC_types,
  ref_map_cell_resolution,
  kernel_radius = 1
)
}
\arguments{
\item{assigned_ref_map}{Data frame of the reference map which should be the
scale at which you want to downscale the land cover data. The data frame
must have a column called `coarse_ID` which assigns each cell in the
reference map to the nearest coarse-scale cell.}

\item{ref_map_LC_types}{Vector of land cover types in the reference map. All
land cover types should be column names in the reference map.}

\item{ref_map_cell_resolution}{Resolution of one cell in the reference map,
in the form `c(x, y)`.}

\item{kernel_radius}{Radius of cells to include in the kernel density
calculation. A value of 1 means that the neighbour cells used to calculate
kernel density will be 1 cell in every direction around the focal cell.
Defaults to 1.}
}
\value{
Data frame with the kernel density value for each land cover type in
  each cell.
}
\description{
Calculates kernel densities for all cell and land cover types in a reference
  map data frame.
}
