% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelDensities.R
\name{calculateKernelDensitiesForOneCell}
\alias{calculateKernelDensitiesForOneCell}
\title{Calculate kernel densities for all land cover types in one cell}
\usage{
calculateKernelDensitiesForOneCell(
  grid_cell,
  assigned_ref_map,
  ref_map_LC_types,
  cell_x_dist,
  cell_y_dist
)
}
\arguments{
\item{grid_cell}{Single row from reference map data frame with x and y values
and the area of each land cover type in that cell.}

\item{assigned_ref_map}{Data frame of the reference map which should be the
scale at which you want to downscale the land cover data. The data frame
must have a column called `coarse_ID` which assigns each cell in the
reference map to the nearest coarse-scale cell.}

\item{ref_map_LC_types}{Vector of land cover types in the reference map. For
an area-based reference map, all land cover types should be column names in
the reference map.}

\item{cell_x_dist}{Radius of the cell neighbourhood in the x-direction.}

\item{cell_y_dist}{Radius of the cell neighbourhood in the y-direction.}
}
\value{
Named vector of the kernel density for each land-use within the
  grid cell.
}
\description{
Calculates the kernel density values for all land cover types in a single
  grid cell.
}
