% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelDensities.R
\name{calculateKernelDensitiesForOneCell}
\alias{calculateKernelDensitiesForOneCell}
\title{Calculate kernel densities for one land cover class in one cell}
\usage{
calculateKernelDensitiesForOneCell(
  grid_cell,
  ref_map_df,
  LC_class,
  kernel_xy_dist
)
}
\arguments{
\item{grid_cell}{Single row from the `ref_map_cells_df` data frame with x-
and y-coordinates and the area of each land cover class in that cell.}

\item{ref_map_df}{Data frame of the reference map for the current timestep.}

\item{LC_class}{Single land cover class for which to calculate kernel
densities. Must be a column name in `ref_map_cells_df` and
`ref_map`.}

\item{kernel_xy_dist}{Vector in the format `c(x, y)` that gives the radius of
the cell neighbourhood in the x- and y- directions.}
}
\value{
The kernel density value for the user-specified land cover class and
  grid cell.
}
\description{
Calculates the kernel density values for one land cover class in a single
  grid cell.
}
