% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverPreProcessing.R
\name{calculateLCDeltas}
\alias{calculateLCDeltas}
\title{Calculate land-use delta values}
\usage{
calculateLCDeltas(LC_map_1, LC_map_2, LC_classes, x_col = "x", y_col = "y")
}
\arguments{
\item{LC_map_1}{Land cover data frame.}

\item{LC_map_2}{Second land cover data frame from the subsequent timestep.}

\item{LC_classes}{Vector of land cover classes, which must be column names in
both data frames.}
}
\value{
A data frame with the delta value for each land cover type between
  the two input timesteps. The function works by subtracting land cover areas
  in `LC_map_1` from those in `LC_map_2`.
}
\description{
Calculates the change in each land cover type (delta) between two timesteps
  of PLUMv2 output.
}
