% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverReconciliation.R
\name{calculateRefMapAreaForCoarseCell}
\alias{calculateRefMapAreaForCoarseCell}
\title{Calculate area of reference map cells assigned to a coarse grid cell}
\usage{
calculateRefMapAreaForCoarseCell(coarse_cell, ref_map_df, ref_map_cell_area)
}
\arguments{
\item{coarse_cell}{A row containing land-use areas and ID number for a single
coarse grid cell.}

\item{ref_map_df}{Data frame of the reference map for the current timestep.}

\item{ref_map_cell_area}{Area of a single reference map grid cell.}
}
\value{
Area of reference map cells that are assigned to the given coarse
  grid cell.
}
\description{
Calculates the area of reference map cells that are assigned to a single
  coarse grid cell.
}
