% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{calculateTentativeConversionValues}
\alias{calculateTentativeConversionValues}
\title{Calculate tentative conversion values}
\usage{
calculateTentativeConversionValues(LC_delta_for_tc, kernel_densities)
}
\arguments{
\item{LC_delta_for_tc}{Change in area of one land cover class (delta) within
one coarse-scale grid cell.}

\item{kernel_densities}{Vector of kernel densities for the land cover class
of interest within reference map cells selected for land cover allocation.}
}
\value{
Vector with a tentative land cover conversion area for each reference
  map cell.
}
\description{
Calculates tentative land cover conversion values based on the change in area
  of one land cover class in a coarse-scale cell and the kernel density
  values of reference map cells selected for land cover allocation.
}
