% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverDownscalingAlgorithm.R
\name{convertDiscreteLCToLCAreas}
\alias{convertDiscreteLCToLCAreas}
\title{Convert a discrete land cover map to one containing the area of each land
  cover class per cell}
\usage{
convertDiscreteLCToLCAreas(
  ref_map_df,
  ref_map_LC_classes,
  ref_map_cell_area,
  LC_column_name
)
}
\arguments{
\item{ref_map_df}{Data frame of the reference map for the current timestep.}

\item{ref_map_LC_classes}{Vector of land cover types in the reference map. For
an area-based reference map, all land cover types should be column names in
the reference map.}

\item{ref_map_cell_area}{Area of a single reference map grid cell.}

\item{LC_column_name}{For a discrete reference map, gives the name of the
column containing the land cover class for each grid cell.}
}
\value{
A data frame of land cover areas per grid cell. The first two columns
  give the x- and y-coordinates of each cell. Remaining columns give the area
  of each land cover class in each cell.
}
\description{
Takes a land cover map with one land cover class per grid cell, and converts
  it to a map with the area of each land cover class per cell.
}
