% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverDownscalingAlgorithm.R
\name{convertDiscreteLCToLCAreasOneCell}
\alias{convertDiscreteLCToLCAreasOneCell}
\title{Convert discrete land cover in one grid cell to a vector with the area of
  land cover classes}
\usage{
convertDiscreteLCToLCAreasOneCell(
  LC_class,
  ref_map_LC_types,
  ref_map_cell_area
)
}
\arguments{
\item{LC_class}{The land cover class in the grid cell.}

\item{ref_map_LC_types}{Vector of land cover types in the reference map. For
an area-based reference map, all land cover types should be column names in
the reference map.}

\item{ref_map_cell_area}{Area of a reference map grid cell.}
}
\value{
A vector with the area of each of the specified land cover classes in
  the given grid cell.
}
\description{
Converts a discrete land cover class for one grid cell to a vector containing
  the area of user-specified land cover classes in that cell.
}
