% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverDownscalingAlgorithm.R
\name{createOutputDir}
\alias{createOutputDir}
\title{Create the output directory for land cover downscaling}
\usage{
createOutputDir(output_dir_path)
}
\arguments{
\item{output_dir_path}{Path to directory in which to save the downscaled
land cover map files. Will be created if it does not already exist.}
}
\value{
Creates an output directory with the path given in `output_dir_path`.
}
\description{
Create the output directory for land cover downscaling
}
