% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverDownscalingAlgorithm.R
\name{downscaleLC}
\alias{downscaleLC}
\title{Downscale land cover data}
\usage{
downscaleLC(
  ref_map_file_name,
  LC_deltas_file_list,
  LC_delta_types,
  LC_deltas_cell_area,
  ref_map_type = "areas",
  LC_column_name = "Land_cover",
  ref_map_LC_types,
  ref_map_cell_area,
  ref_map_cell_resolution,
  final_LC_types,
  kernel_radius,
  transition_priorities,
  intensification_ratio,
  output_file_prefix,
  output_dir_path
)
}
\arguments{
\item{ref_map_file_name}{Path of file name for the reference map for
downscaling. The reference map file should be in tab-separated format with
the extension `.txt`. The resolution of the reference map will determine
the resolution of the output downscaled land cover maps.}

\item{LC_deltas_file_list}{List of file names for land cover change (delta).
The algorithm expects one land cover change file per time step. All land
cover change files must be in tab-separated format with the `.txt` file
extension. The list must be ordered by time, with the first time step being
the first file in the list, the second time step the second file, and so
on. Land cover change files can be generated using the `calculateLCDeltas`
function from this package.}

\item{LC_delta_types}{Column names of land cover types in the land cover
change data frame `LC_deltas`.}

\item{LC_deltas_cell_area}{Area of a coarse-scale grid cell.}

\item{ref_map_type}{Specifies whether the reference map is discrete (contains
one land cover class per cell) or area-based (provides the area of each
land cover class in each cell). Must be one of "areas" or "discrete".}

\item{LC_column_name}{For a discrete reference map, gives the name of the
column containing the land cover class for each grid cell.}

\item{ref_map_LC_types}{Vector of land cover types in the reference map. For
an area-based reference map, all land cover types should be column names in
the reference map.}

\item{ref_map_cell_area}{Area of a reference map grid cell.}

\item{ref_map_cell_resolution}{Resolution of one cell in the reference map,
in the form `c(x, y)`.}

\item{final_LC_types}{A matrix containing the fraction of each coarse-scale
land cover type that contributes to each reference map land cover type.
Columns should contain reference map land cover types, and rows are the
coarse-scale land cover types. Each cell should contain the proportion of
the coarse-scale land cover type that contributes to the fine-scale type in
the output map.}

\item{kernel_radius}{Radius of cells to include in the kernel density
calculation. A value of 1 means that the neighbour cells used to calculate
kernel density will be 1 cell in every direction around the focal cell.
Defaults to 1.}

\item{transition_priorities}{Matrix containing transition priorities for land
cover allocation. Each row of the matrix should give the order in which one
land cover type is converted to others.}

\item{intensification_ratio}{Ratio of land intensification versus land
expansion for the allocation algorithm.}

\item{output_file_prefix}{Prefix for output downscaled land cover map files.}

\item{output_dir_path}{Path to directory in which to saved the downscaled
land cover map files.}
}
\value{
Tab-separated text files containing downscaled land cover maps, with
  one file per input land cover change file. In each output file, the first
  two columns are x and y coordinates for the map. Subsequent columns contain
  land cover for each cell. The 'ref_ID' column gives an identification
  number for each grid cell, and the 'coarse_ID' column specifies the
  coarse-scale cell to which each fine-scale cell was assigned to during
  downscaling.
}
\description{
Downscales land cover by applying land cover change from coarse-scale input
  maps to a fine-scale reference map. The method used for downscaling land
  cover is based on that in West et al. (2014) and Le Page et al. (2016).
  Briefly, all fine-scale reference map cells are assigned to a single
  coarse-scale cell using a nearest neighbour approach. Kernel density values
  are calculated for each fine-scale cell and land cover type based on the
  amount of each land cover type in neighbouring cells. The kernel density
  values are used during the land cover allocation process to determine how
  much of a land cover type should be allocated to each grid cell; cells with
  higher kernel density values will be allocated more land cover. Land cover
  change is allocated in three steps: intensification, expansion, followed by
  a second round of intensification.
}
\details{
Land cover allocation is performed in three steps: a first round of
  intensification, where land cover is allocated to cells where it already
  occurs; a round of expansion, where land cover is allocated to cells where
  it does not exist; and a final round of intensification to ensure that all
  land cover change has been allocated. The `intensification_ratio` parameter
  specifies how much land cover the algorithm should attempt to allocate via
  intensification versus expansion. An intensification ratio of 0.8 gives a
  target of 80\% of land cover to be allocated via intensification and 20\%
  by expansion.
}
