% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{getCellsForAllocation}
\alias{getCellsForAllocation}
\title{Get a data frame of candidate fine-scale cells for land cover allocation}
\usage{
getCellsForAllocation(ref_map, LC_from_name)
}
\arguments{
\item{ref_map}{'SpatRaster' object with fine resolution cells.}

\item{LC_from_name}{Name of the land cover class that is being converted to
\code{LC_to_name}.}
}
\value{
Data frame of fine-scale cells that contain \code{LC_from_name}.
}
\description{
Fine-scale cells are selected for allocation if they contain the land cover
class that is being converted to a different class.
}
