% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{getCellsForAllocationKdZero}
\alias{getCellsForAllocationKdZero}
\title{Return a data frame of fine-scale cells with kernel density equal to zero}
\usage{
getCellsForAllocationKdZero(cells_for_allocation)
}
\arguments{
\item{cells_for_allocation}{Data frame of fine-scale cells selected to
receive a given land cover class.}
}
\value{
Data frame of fine-scale cells with kernel density equal to 0 that
  have been randomly sorted.
}
\description{
Return a data frame of fine-scale cells with kernel density equal to zero
}
