% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{getCellsToExpandLC}
\alias{getCellsToExpandLC}
\title{Get cells in which to expand one land cover type}
\usage{
getCellsToExpandLC(ref_map_df_with_IDs, LC_inc_name, LC_dec_name, coarse_ID)
}
\arguments{
\item{ref_map_df_with_IDs}{Reference map data frame with a column containing
the ID of the coarse grid cell to which each reference map cell was
assigned to. This is output from the `assignFineScaleCells` function.}

\item{LC_inc_name}{Name of the land cover that is increasing.}

\item{LC_dec_name}{Name of the land cover that `LC_inc_name` is being
converted to.}

\item{coarse_ID}{Identification number of the coarse-scale cell within which
`LC_inc_name` has increased.}
}
\value{
Data frame of cells from the reference map that meet the criteria for
  expansion of one land cover type into another.
}
\description{
Find cells which match the conditions for expanding one land cover type into
  a second. Cells meet the criteria for expansion if the increasing land
  cover is 0 and the decreasing land cover is greater than 0 within the cell.
}
