% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{getCellsToIntensifyLC}
\alias{getCellsToIntensifyLC}
\title{Get cells in which to intensify one land cover class}
\usage{
getCellsToIntensifyLC(ref_map_df, LC_to_name, LC_from_name, coarse_ID)
}
\arguments{
\item{ref_map_df}{Data frame of the reference map for the current timestep.}

\item{LC_to_name}{Name of the land cover class that is increasing.}

\item{LC_from_name}{Name of the land cover class that is being converted to
`LC_to_name`.}

\item{coarse_ID}{Identification number of the coarse-scale cell within which
`LC_to_name` has increased.}
}
\value{
Data frame of cells from the reference map that meet the criteria for
  intensification of one land cover class into another.
}
\description{
Find cells which match the conditions for intensifying one land cover class
  into a second. Cells meet the criteria for intensification if both land
  cover classes are greater than 0 within the cell.
}
