% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{getCellsToIntensifyLC}
\alias{getCellsToIntensifyLC}
\title{Get cells in which to intensify one land cover type}
\usage{
getCellsToIntensifyLC(
  updated_ref_map_df_with_IDs,
  LC_inc_name,
  LC_dec_name,
  coarse_ID
)
}
\arguments{
\item{updated_ref_map_df_with_IDs}{Reference map data frame with a column
containing the ID of the coarse grid cell to which each reference map cell
was assigned to.}

\item{LC_inc_name}{Name of the land cover that is increasing.}

\item{LC_dec_name}{Name of the land cover that `LC_inc_name` is being
converted to.}

\item{coarse_ID}{Identification number of the coarse-scale cell within which
`LC_inc_name` has increased.}
}
\value{
Data frame of cells from the reference map that meet the criteria for
  intensification of one land cover type into another.
}
\description{
Find cells which match the conditions for intensifying one land cover type
  into a second. Cells meet the criteria for intensification if both land
  cover types are greater than 0 within the cell.
}
