% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{getLCConversions}
\alias{getLCConversions}
\title{Get amount of one land cover type converted into another}
\usage{
getLCConversions(
  cells_for_allocation,
  assigned_ref_map,
  ref_map_cell_resolution,
  kernel_radius,
  LC_to_name,
  LC_from_name,
  LC_to_delta,
  LC_from_delta
)
}
\arguments{
\item{cells_for_allocation}{Data frame of reference cells to which the
increasing land cover type will be allocated.}

\item{assigned_ref_map}{Data frame of the reference map which should be the
scale at which you want to downscale the land cover data. The data frame
must have a column called `coarse_ID` which assigns each cell in the
reference map to the nearest coarse-scale cell.}

\item{ref_map_cell_resolution}{Resolution of one cell in the reference map,
in the form `c(x, y)`.}

\item{kernel_radius}{Radius of cells to include in the kernel density
calculation. A value of 1 means that the neighbour cells used to calculate
kernel density will be 1 cell in every direction around the focal cell.
Defaults to 1.}

\item{LC_to_name}{Name of the land cover that is increasing.}

\item{LC_from_name}{Name of the land cover that is being converted to
`LC_to_name`.}

\item{LC_to_delta}{The land cover change amount (delta) for the increasing
land cover type, which is equivalent to the amount that this land cover
type increased in the coarse-scale cell during one timestep.}

\item{LC_from_delta}{The land cover delta value for the decreasing land cover
type, which is the amount that this land cover type decreased in the
coarse-scale cell during one timestep.}
}
\value{
Two column data frame. First column contains IDs of reference map
  cells, second column is the amount of the increasing land cover type
  allocated to each cell.
}
\description{
Calculates the amount of one land cover type that is converted into a second
  land cover type for each reference map cell assigned to a coarse-scale
  cell.
}
