% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{getLCConversions}
\alias{getLCConversions}
\title{Get amount of one land cover type converted into another}
\usage{
getLCConversions(
  cells_for_allocation,
  ref_map_df,
  ref_map_cell_resolution,
  kernel_radius,
  LC_to_name,
  LC_from_name,
  LC_to_delta,
  LC_from_delta
)
}
\arguments{
\item{cells_for_allocation}{Data frame of reference map cells to which the
increasing land cover class will be allocated.}

\item{ref_map_df}{Data frame of the reference map for the current timestep.}

\item{ref_map_cell_resolution}{Resolution of one cell in the reference map,
in the form `c(x, y)`.}

\item{kernel_radius}{Radius of cells to include in the kernel density
calculation. A value of 1 means that the neighbour cells used to calculate
kernel density will be 1 cell in every direction around the focal cell.
Defaults to 1.}

\item{LC_to_name}{Name of the land cover class that is increasing.}

\item{LC_from_name}{Name of the land cover class that is being converted to
`LC_to_name`.}

\item{LC_to_delta}{The area by which the increasing land cover class grew in
the coarse-scale cell the current timestep.}

\item{LC_from_delta}{The area by which the decreasing land cover class
contracted in the coarse-scale cell in the current timestep.}
}
\value{
Two column data frame. First column contains IDs of reference map
  cells, second column is the amount of the increasing land cover class
  allocated to each cell.
}
\description{
Calculates the amount of one land cover type that is converted into a second
  land cover type for each reference map cell assigned to a coarse-scale
  cell.
}
