% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{getLCDeltaForTentativeConversion}
\alias{getLCDeltaForTentativeConversion}
\title{Get land cover delta area for tentative land cover conversions}
\usage{
getLCDeltaForTentativeConversion(LC_to_delta, LC_from_delta)
}
\arguments{
\item{LC_to_delta}{The area by which the increasing land cover class grew in
the coarse-scale cell the current timestep.}

\item{LC_from_delta}{The area by which the decreasing land cover class
contracted in the coarse-scale cell in the current timestep.}
}
\value{
Value of land cover delta to be used in tentative land cover
  conversions.
}
\description{
Returns land cover change area (delta) to be used in tentative land cover
  conversions, which is the minimum of the absolute increasing and decreasing
  land cover class delta values.
}
