% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{getLCTransitions}
\alias{getLCTransitions}
\title{Get the transition matrix for one cell}
\usage{
getLCTransitions(LC_deltas_cell, LC_classes)
}
\arguments{
\item{LC_deltas_cell}{One row of a land cover deltas data frame, which
corresponds to one coarse-scale grid cell.}

\item{LC_classes}{Vector of land cover classes in the coarse-scale grid cell.}
}
\value{
Data frame giving order and magniture of land cover transitions.
}
\description{
The transition matrix determines the area of land cover change that is
  converted to each land cover class.
}
