% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelDensities.R
\name{kernelDensityFunction}
\alias{kernelDensityFunction}
\title{Calculate kernel density value for a single land cover in a single cell}
\usage{
kernelDensityFunction(LC_values, distance_values, number_of_neighbour_cells)
}
\arguments{
\item{LC_values}{Vector of land cover areas within each neighbour cell.}

\item{distance_values}{Vector giving the distance of each neighbour cell
from the focal cell, should be the same length as `LC_values`.}

\item{number_of_neighbour_cells}{Number of neighbour cells.}
}
\value{
The kernel density of the given land-use in the focal cell.
}
\description{
Calculates the kernel density value for a single land cover in a single
  fine-scale grid cell.
}
