% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAreaChecks.R
\name{LCAllocationChecks}
\alias{LCAllocationChecks}
\title{Check unallocated land cover change and the sum of land cover areas in each
  reference map cell}
\usage{
LCAllocationChecks(
  LC_deltas_df,
  LC_deltas_cell_area,
  LC_classes,
  ref_map_df,
  ref_map_cell_area
)
}
\description{
Check unallocated land cover change and the sum of land cover areas in each
  reference map cell
}
