% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{LCAllocationChecks}
\alias{LCAllocationChecks}
\title{Check unallocated land cover change and the sum of land cover areas in each
  reference map cell}
\usage{
LCAllocationChecks(LC_deltas_df, ref_map_df, LC_classes, ref_map_cell_area)
}
\arguments{
\item{LC_deltas_df}{Data frame of coarse-scale land cover change areas for
the current timestep.}

\item{ref_map_df}{Data frame of the reference map for the current timestep.}

\item{LC_classes}{Vector of land cover classes in the coarse-scale grid cell.}

\item{ref_map_cell_area}{Area of a single reference map grid cell.}
}
\value{
Returns nothing. Checks for unallocated land cover change and that
  the sum of land cover areas in each reference map cell is equal to the
  original area of the cell.
}
\description{
Check unallocated land cover change and the sum of land cover areas in each
  reference map cell
}
