% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverReconciliation.R
\name{processLCDeltas}
\alias{processLCDeltas}
\title{Land cover change processing}
\usage{
processLCDeltas(
  LC_deltas,
  ref_map_df,
  coarse_cell_area,
  ref_map_cell_area,
  final_LC_types
)
}
\arguments{
\item{LC_deltas}{Data frame of land cover changes between two coarse-scale
timesteps. Must be the same projection as `ref_map_df`.}

\item{ref_map_df}{Data frame of the reference map, which is the scale at
which you want to downscale the land cover data.}

\item{coarse_cell_area}{Area of a coarse-scale grid cell.}

\item{ref_map_cell_area}{Area of a reference map grid cell.}

\item{final_LC_types}{A matrix containing the fraction of each coarse-scale
land cover type that contributes to each reference map land cover type.
Columns should contain reference map land cover types, and rows are the
coarse-scale land cover types. Each cell should contain the proportion of
the coarse-scale land cover type that contributes to the fine-scale type in
the output map.}
}
\value{
ReconciledLandUses class containing a data frame of reference map
  cells assigned to coarse-scale cells, and a second data frame of
  coarse-scale cells with adjusted and aggregated land-use areas.
}
\description{
Performs the following actions: assigns fine-scale cells to coarse-scale
  cells, reconciles the land cover change values for each coarse-scale cell
  with fine-scale areas, and aggregates land cover types.
}
