% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landUseReconciliation.R
\name{reconcileLandUses}
\alias{reconcileLandUses}
\title{Land-use reconciliation}
\usage{
reconcileLandUses(
  coarse_scale_raster,
  fine_scale_raster,
  coarse_cell_area,
  fine_cell_area,
  final_land_use_types
)
}
\arguments{
\item{coarse_scale_raster}{A raster of land-use areas at a coarse-scale, i.e.
output from PLUMv2.}

\item{fine_scale_raster}{A raster of land-use areas at the scale to which you
want to downscale the `coarse_scale_raster`. Must be in the same projection
as the `coarse_scale_raster`.}

\item{coarse_cell_area}{The area of a grid cell in the
`coarse_scale_raster`.}

\item{fine_cell_area}{The area of a grid cell in the `fine_scale_raster`.}

\item{final_land_use_types}{A matrix containing the fraction of each
coarse-scale land-use type that contributes to each fine-scale land-use
type. Columns should contain fine-scale land-use categories, and rows are
the coarse-scale land-use categories. Each cell should contain the
proportion of the coarse-scale land-use category that contributes to the
fine-scale category in the output map.}
}
\value{
A ReconciledLandUses class containing a data frame of fine-scale
  cells assigned to coarse-scale cells, and a second data frame of
  coarse-scale cells with adjusted and aggregated land-use areas.
}
\description{
Performs the following actions: assigns fine-scale cells to coarse-scale
  cells, reconciles the coarse-scale areas with fine-scale areas, and
  aggregates land-uses.
}
