% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverReconciliation.R
\name{reconcileLCDeltas}
\alias{reconcileLCDeltas}
\title{Reconcile land cover deltas}
\usage{
reconcileLCDeltas(
  LC_deltas,
  assigned_ref_map,
  LC_delta_types,
  LC_deltas_cell_area,
  ref_map_cell_area
)
}
\arguments{
\item{LC_deltas}{Data frame of land cover changes between two coarse-scale
timesteps. Must be the same projection as `assigned_ref_map`.}

\item{assigned_ref_map}{Data frame of the reference map which should be the
scale at which you want to downscale the land cover data. The data frame
must have a column called `coarse_ID` which assigns each cell in the
reference map to the nearest coarse-scale cell.}

\item{LC_delta_types}{Column names of land cover types in the land cover
change data frame `LC_deltas`.}

\item{LC_deltas_cell_area}{Area of a coarse-scale grid cell.}

\item{ref_map_cell_area}{Area of a reference map grid cell.}
}
\value{
A data frame of land-use areas within coarse-scale cells, with every
  area adjusted for the area of fine-scale cells associated with that coarse
  grid cell.
}
\description{
Reconcile land cover deltas for each coarse-scale grid cell with the area of
  reference map cells assigned to that grid cell.
}
\details{
This function uses equation 1 from the Le Page et al. (2016) paper to
  reconcile coarse- and fine-scale land cover areas.
}
