% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landUseReconciliation.R
\name{reconcileLCDeltas}
\alias{reconcileLCDeltas}
\title{Reconcile land cover deltas}
\usage{
reconcileLCDeltas(
  LC_deltas,
  ref_map_df_with_IDs,
  coarse_LC_types,
  coarse_cell_area,
  ref_map_cell_area
)
}
\arguments{
\item{LC_deltas}{Data frame of land cover changes between two coarse-scale
timesteps. Must be the same projection as `ref_map_df`.}

\item{ref_map_df_with_IDs}{Reference map data frame with a column containing
the ID of the coarse grid cell to which each reference map cell was
assigned to. This is output from the `assignFineScaleCells` function.}

\item{coarse_LC_types}{Vector of land cover types in the land cover deltas
data frame.}

\item{coarse_cell_area}{Area of a coarse-scale grid cell.}

\item{ref_map_cell_area}{Area of a reference map grid cell.}
}
\value{
A data frame of land-use areas within coarse-scale cells, with every
  area adjusted for the area of fine-scale cells associated with that coarse
  grid cell.
}
\description{
Reconcile land cover deltas for each coarse-scale grid cell with the area of
  reference map cells assigned to that grid cell.
}
\details{
This function uses equation 1 from the Le Page et al. (2016) paper to
  reconcile coarse- and fine-scale land cover areas.
}
