% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{runLCAllocation}
\alias{runLCAllocation}
\title{Run allocation of land covers to reference map by intensification and
  expansion}
\usage{
runLCAllocation(LC_allocation_params)
}
\arguments{
\item{LC_allocation_params}{`LCAllocationParams` object containing all the
parameters required to allocate land cover change to a reference map for a
single timestep.}
}
\value{
New land cover data frame at the same resolution as the given
  reference map with the specified land cover change areas allocated.
}
\description{
This function will run the allocation of all land cover change (delta) values
  from a coarse-scale map, such as the output from one timestep of a global
  land-use model, to a fine-scale reference map. The method employed is from
  West et al. (2014) and Le Page et al. (2016). Land cover change is
  allocated in three steps: a first round of intensification, meaning land
  cover is allocated to cells where it already occurs; a round of expansion,
  where land cover is allocated to cells where it does not exist; and a final
  round of intensification to make sure all land cover change has been
  allocated.
}
\details{
The amount of land cover change that is allocated by intensification versus
  expansion is determined by the `intensification_ratio` parameter. An
  intensification ratio of 0.8 gives a target of 80 percent of land cover to
  be allocated via intensification and 20 percent by expansion.

  The amount of land cover change for a single land cover type that is
  allocated to a fine-scale reference cell is currently determined by the
  kernel density value of that cell. A kernel density value is a relative
  measure of the amount of a land cover type in neighbouring cells.
}
