% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{sortCellsForAllocation}
\alias{sortCellsForAllocation}
\title{Sort cells before land cover allocation}
\usage{
sortCellsForAllocation(cells_for_allocation)
}
\arguments{
\item{cells_for_allocation}{Data frame of fine-scale cells selected to
receive a given land cover class.}
}
\value{
Data frame of fine-scale cells ordered from highest to lowest kernel
density. Cells with a kernel density value of zero are randomly sorted at
the end of the data frame.
}
\description{
Sort cells before land cover allocation
}
