% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{sumLCDeltas}
\alias{sumLCDeltas}
\title{Sum land cover change values}
\usage{
sumLCDeltas(LC_deltas_one, LC_deltas_two)
}
\arguments{
\item{LC_deltas_one}{An `LCMap` object containing land cover change areas.}

\item{LC_deltas_two}{A second `LCMap` object containing land cover change
areas.}
}
\value{
`LCMap` object containing the sum of land cover change from the two
  input objects.
}
\description{
Sums the land cover change values in two data frames. The land cover types
  and grid cells must be the same in each data frame.
}
