% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landCoverAllocation.R
\name{updateOneRefMapCellWithLCConversions}
\alias{updateOneRefMapCellWithLCConversions}
\title{Update one reference map cell with land cover conversion areas}
\usage{
updateOneRefMapCellWithLCConversions(
  ref_map_cell,
  LC_to_name,
  LC_from_name,
  actual_conversion
)
}
\arguments{
\item{ref_map_cell}{One row from a reference map data frame, which is
equivalent to one grid cell.}

\item{LC_to_name}{Name of the land cover that is increasing.}

\item{LC_from_name}{Name of the land cover that is being converted to
`LC_to_name`.}

\item{actual_conversion}{The actual area of land cover to be converted from
the decreasing land cover (`LC_from_name`) to the increasing land cover
(`LC_to_name`) in this reference map cell.}
}
\value{
A one row data frame with updated land cover areas.
}
\description{
Updates one cell from a reference map with land cover conversion areas from
  the conversion of one land cover type to another.
}
